+++ janitor/polars/__init__.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+from typing import Any
+
 from polars.type_aliases import ColumnNameOrSelector
 
 from janitor.utils import import_message
@@ -25,71 +27,215 @@ class PolarsFrame:
     def complete(
         self,
         *columns: ColumnNameOrSelector,
-        fill_value: dict | float | int | str = None,
+        fill_value: dict | Any | pl.Expr = None,
         explicit: bool = True,
         sort: bool = False,
+        by: ColumnNameOrSelector = None,
     ) -> pl.DataFrame:
         """
-        Complete a data frame with missing combinations of data.
+        Turns implicit missing values into explicit missing values
 
         It is modeled after tidyr's `complete` function.
         In a way, it is the inverse of `pl.drop_nulls`,
         as it exposes implicitly missing rows.
 
-        Combinations of column names or a list/tuple of column names, or even a
-        dictionary of column names and new values are possible.
-        If a dictionary is passed,
-        the user is required to ensure that the values are unique 1-D arrays.
-        The keys in a dictionary must be present in the DataFrame.
+        If the combination involves multiple columns, pass it as a struct.
+        If new values need to be introduced, a polars Expression
+        with the new values can be passed, as long as the polars Expression
+        has a name that already exists in the DataFrame.
 
         Examples:
             >>> import polars as pl
             >>> import janitor.polars
+            >>> pl.Config.set_tbl_cols(-1)
+            >>> pl.Config.set_tbl_rows(-1)
+            >>> df = pl.DataFrame(
+            ...     dict(
+            ...         group=(1, 2, 1, 2),
+            ...         item_id=(1, 2, 2, 3),
+            ...         item_name=("a", "a", "b", "b"),
+            ...         value1=(1, np.nan, 3, 4),
+            ...         value2=range(4, 8),
+            ...     )
+            ... )
+            >>> df
+            shape: (4, 5)
+            ┌───────┬─────────┬───────────┬────────┬────────┐
+            │ group ┆ item_id ┆ item_name ┆ value1 ┆ value2 │
+            │ ---   ┆ ---     ┆ ---       ┆ ---    ┆ ---    │
+            │ i64   ┆ i64     ┆ str       ┆ i64    ┆ i64    │
+            ╞═══════╪═════════╪═══════════╪════════╪════════╡
+            │ 1     ┆ 1       ┆ a         ┆ 1      ┆ 4      │
+            │ 2     ┆ 2       ┆ a         ┆ null   ┆ 5      │
+            │ 1     ┆ 2       ┆ b         ┆ 3      ┆ 6      │
+            │ 2     ┆ 3       ┆ b         ┆ 4      ┆ 7      │
+            └───────┴─────────┴───────────┴────────┴────────┘
+
+            Generate all possible combinations of
+            `group`, `item_id`, and `item_name`
+            (whether or not they appear in the data)
+            >>> df.janitor.complete("group", "item_id", "item_name", sort=True)
+            shape: (12, 5)
+            ┌───────┬─────────┬───────────┬────────┬────────┐
+            │ group ┆ item_id ┆ item_name ┆ value1 ┆ value2 │
+            │ ---   ┆ ---     ┆ ---       ┆ ---    ┆ ---    │
+            │ i64   ┆ i64     ┆ str       ┆ i64    ┆ i64    │
+            ╞═══════╪═════════╪═══════════╪════════╪════════╡
+            │ 1     ┆ 1       ┆ a         ┆ 1      ┆ 4      │
+            │ 1     ┆ 1       ┆ b         ┆ null   ┆ null   │
+            │ 1     ┆ 2       ┆ a         ┆ null   ┆ null   │
+            │ 1     ┆ 2       ┆ b         ┆ 3      ┆ 6      │
+            │ 1     ┆ 3       ┆ a         ┆ null   ┆ null   │
+            │ 1     ┆ 3       ┆ b         ┆ null   ┆ null   │
+            │ 2     ┆ 1       ┆ a         ┆ null   ┆ null   │
+            │ 2     ┆ 1       ┆ b         ┆ null   ┆ null   │
+            │ 2     ┆ 2       ┆ a         ┆ null   ┆ 5      │
+            │ 2     ┆ 2       ┆ b         ┆ null   ┆ null   │
+            │ 2     ┆ 3       ┆ a         ┆ null   ┆ null   │
+            │ 2     ┆ 3       ┆ b         ┆ 4      ┆ 7      │
+            └───────┴─────────┴───────────┴────────┴────────┘
+
+            Cross all possible `group` values with the unique pairs of
+            `(item_id, item_name)` that already exist in the data.
+            For such situations, where there is a group of columns,
+            pass it in as a struct:
+            >>> df.janitor.complete("group", pl.struct("item_id", "item_name"), sort=True)
+            shape: (8, 5)
+            ┌───────┬─────────┬───────────┬────────┬────────┐
+            │ group ┆ item_id ┆ item_name ┆ value1 ┆ value2 │
+            │ ---   ┆ ---     ┆ ---       ┆ ---    ┆ ---    │
+            │ i64   ┆ i64     ┆ str       ┆ i64    ┆ i64    │
+            ╞═══════╪═════════╪═══════════╪════════╪════════╡
+            │ 1     ┆ 1       ┆ a         ┆ 1      ┆ 4      │
+            │ 1     ┆ 2       ┆ a         ┆ null   ┆ null   │
+            │ 1     ┆ 2       ┆ b         ┆ 3      ┆ 6      │
+            │ 1     ┆ 3       ┆ b         ┆ null   ┆ null   │
+            │ 2     ┆ 1       ┆ a         ┆ null   ┆ null   │
+            │ 2     ┆ 2       ┆ a         ┆ null   ┆ 5      │
+            │ 2     ┆ 2       ┆ b         ┆ null   ┆ null   │
+            │ 2     ┆ 3       ┆ b         ┆ 4      ┆ 7      │
+            └───────┴─────────┴───────────┴────────┴────────┘
+
+            Fill in nulls:
+            >>> df.janitor.complete(
+            ...     "group",
+            ...     pl.struct("item_id", "item_name"),
+            ...     fill_value={"value1": 0, "value2": 99},
+            ...     explicit=True,
+            ...     sort=True,
+            ... )
+            shape: (8, 5)
+            ┌───────┬─────────┬───────────┬────────┬────────┐
+            │ group ┆ item_id ┆ item_name ┆ value1 ┆ value2 │
+            │ ---   ┆ ---     ┆ ---       ┆ ---    ┆ ---    │
+            │ i64   ┆ i64     ┆ str       ┆ i64    ┆ i64    │
+            ╞═══════╪═════════╪═══════════╪════════╪════════╡
+            │ 1     ┆ 1       ┆ a         ┆ 1      ┆ 4      │
+            │ 1     ┆ 2       ┆ a         ┆ 0      ┆ 99     │
+            │ 1     ┆ 2       ┆ b         ┆ 3      ┆ 6      │
+            │ 1     ┆ 3       ┆ b         ┆ 0      ┆ 99     │
+            │ 2     ┆ 1       ┆ a         ┆ 0      ┆ 99     │
+            │ 2     ┆ 2       ┆ a         ┆ 0      ┆ 5      │
+            │ 2     ┆ 2       ┆ b         ┆ 0      ┆ 99     │
+            │ 2     ┆ 3       ┆ b         ┆ 4      ┆ 7      │
+            └───────┴─────────┴───────────┴────────┴────────┘
+
+            Limit the fill to only the newly created
+            missing values with `explicit = FALSE`
+            >>> df.janitor.complete(
+            ...     "group",
+            ...     pl.struct("item_id", "item_name"),
+            ...     fill_value={"value1": 0, "value2": 99},
+            ...     explicit=False,
+            ...     sort=True,
+            ... )
+            shape: (8, 5)
+            ┌───────┬─────────┬───────────┬────────┬────────┐
+            │ group ┆ item_id ┆ item_name ┆ value1 ┆ value2 │
+            │ ---   ┆ ---     ┆ ---       ┆ ---    ┆ ---    │
+            │ i64   ┆ i64     ┆ str       ┆ i64    ┆ i64    │
+            ╞═══════╪═════════╪═══════════╪════════╪════════╡
+            │ 1     ┆ 1       ┆ a         ┆ 1      ┆ 4      │
+            │ 1     ┆ 2       ┆ a         ┆ 0      ┆ 99     │
+            │ 1     ┆ 2       ┆ b         ┆ 3      ┆ 6      │
+            │ 1     ┆ 3       ┆ b         ┆ 0      ┆ 99     │
+            │ 2     ┆ 1       ┆ a         ┆ 0      ┆ 99     │
+            │ 2     ┆ 2       ┆ a         ┆ null   ┆ 5      │
+            │ 2     ┆ 2       ┆ b         ┆ 0      ┆ 99     │
+            │ 2     ┆ 3       ┆ b         ┆ 4      ┆ 7      │
+            └───────┴─────────┴───────────┴────────┴────────┘
+
             >>> df = pl.DataFrame(
             ...     {
-            ...         "Aloha": range(3),
-            ...         "Bell Chart": range(3),
-            ...         "Animals@#$%^": range(3)
+            ...         "Year": [1999, 2000, 2004, 1999, 2004],
+            ...         "Taxon": [
+            ...             "Saccharina",
+            ...             "Saccharina",
+            ...             "Saccharina",
+            ...             "Agarum",
+            ...             "Agarum",
+            ...         ],
+            ...         "Abundance": [4, 5, 2, 1, 8],
             ...     }
             ... )
             >>> df
-            shape: (3, 3)
-            ┌───────┬────────────┬──────────────┐
-            │ Aloha ┆ Bell Chart ┆ Animals@#$%^ │
-            │ ---   ┆ ---        ┆ ---          │
-            │ i64   ┆ i64        ┆ i64          │
-            ╞═══════╪════════════╪══════════════╡
-            │ 0     ┆ 0          ┆ 0            │
-            │ 1     ┆ 1          ┆ 1            │
-            │ 2     ┆ 2          ┆ 2            │
-            └───────┴────────────┴──────────────┘
-            >>> df.janitor.clean_names(remove_special=True)
-            shape: (3, 3)
-            ┌───────┬────────────┬─────────┐
-            │ aloha ┆ bell_chart ┆ animals │
-            │ ---   ┆ ---        ┆ ---     │
-            │ i64   ┆ i64        ┆ i64     │
-            ╞═══════╪════════════╪═════════╡
-            │ 0     ┆ 0          ┆ 0       │
-            │ 1     ┆ 1          ┆ 1       │
-            │ 2     ┆ 2          ┆ 2       │
-            └───────┴────────────┴─────────┘
+            shape: (5, 3)
+            ┌──────┬────────────┬───────────┐
+            │ Year ┆ Taxon      ┆ Abundance │
+            │ ---  ┆ ---        ┆ ---       │
+            │ i64  ┆ str        ┆ i64       │
+            ╞══════╪════════════╪═══════════╡
+            │ 1999 ┆ Saccharina ┆ 4         │
+            │ 2000 ┆ Saccharina ┆ 5         │
+            │ 2004 ┆ Saccharina ┆ 2         │
+            │ 1999 ┆ Agarum     ┆ 1         │
+            │ 2004 ┆ Agarum     ┆ 8         │
+            └──────┴────────────┴───────────┘
+
+            Expose missing years from 1999 to 2004 -
+            pass a polars expresion with the new dates,
+            and ensure the expression's name already exists
+            in the DataFrame:
+            >>> expression = pl.int_range(1999,2005).alias('Year')
+            >>> df.janitor.complete(expression,'Taxon',sort=True)
+            shape: (12, 3)
+            ┌──────┬────────────┬───────────┐
+            │ Year ┆ Taxon      ┆ Abundance │
+            │ ---  ┆ ---        ┆ ---       │
+            │ i64  ┆ str        ┆ i64       │
+            ╞══════╪════════════╪═══════════╡
+            │ 1999 ┆ Agarum     ┆ 1         │
+            │ 1999 ┆ Saccharina ┆ 4         │
+            │ 2000 ┆ Agarum     ┆ null      │
+            │ 2000 ┆ Saccharina ┆ 5         │
+            │ 2001 ┆ Agarum     ┆ null      │
+            │ 2001 ┆ Saccharina ┆ null      │
+            │ 2002 ┆ Agarum     ┆ null      │
+            │ 2002 ┆ Saccharina ┆ null      │
+            │ 2003 ┆ Agarum     ┆ null      │
+            │ 2003 ┆ Saccharina ┆ null      │
+            │ 2004 ┆ Agarum     ┆ 8         │
+            │ 2004 ┆ Saccharina ┆ 2         │
+            └──────┴────────────┴───────────┘
 
         !!! info "New in version 0.28.0"
 
         Args:
             *columns: This refers to the columns to be completed.
-                It could be column labels (string type),
-                a list/tuple of column labels, or a dictionary that pairs
-                column labels with new values.
-            fill_value: Scalar value to use instead of nulls
+                It can be a string or a column selector or a polars expression.
+                A polars expression can be used to introduced new values,
+                as long as the polars expression has a name that already exists
+                in the DataFrame.
+            fill_value: Scalar value or polars expression to use instead of nulls
                 for missing combinations. A dictionary, mapping columns names
                 to a scalar value is also accepted.
             explicit: Determines if only implicitly missing values
-                should be filled (`False`), or all nulls existing in the dataframe
+                should be filled (`False`), or all nulls existing in the LazyFrame
                 (`True`). `explicit` is applicable only
                 if `fill_value` is not `None`.
-            sort: Sort DataFrame based on *columns.
+            sort: Sort the DataFrame based on *columns.
+            by: Column(s) to group by.
+                The explicit missing rows are returned per group.
         Returns:
             A polars DataFrame.
         """  # noqa: E501
@@ -99,6 +245,7 @@ class PolarsFrame:
             fill_value=fill_value,
             explicit=explicit,
             sort=sort,
+            by=by,
         )
 
 
@@ -110,73 +257,164 @@ class PolarsLazyFrame:
     def complete(
         self,
         *columns: ColumnNameOrSelector,
-        fill_value: dict | float | int | str = None,
+        fill_value: dict | Any | pl.Expr = None,
         explicit: bool = True,
         sort: bool = False,
+        by: ColumnNameOrSelector = None,
     ) -> pl.DataFrame:
         """
-        Complete a data frame with missing combinations of data.
+        Turns implicit missing values into explicit missing values
 
         It is modeled after tidyr's `complete` function.
         In a way, it is the inverse of `pl.drop_nulls`,
         as it exposes implicitly missing rows.
 
-        Combinations of column names or a list/tuple of column names, or even a
-        dictionary of column names and new values are possible.
-        If a dictionary is passed,
-        the user is required to ensure that the values are unique.
-        The keys in a dictionary must be present in the DataFrame.
+        If the combination involves multiple columns, pass it as a struct.
+        If new values need to be introduced, a polars Expression
+        with the new values can be passed, as long as the polars Expression
+        has a name that already exists in the LazyFrame.
 
         Examples:
             >>> import polars as pl
             >>> import janitor.polars
-            >>> df = pl.DataFrame(
-            ...     {
-            ...         "Aloha": range(3),
-            ...         "Bell Chart": range(3),
-            ...         "Animals@#$%^": range(3)
-            ...     }
+            >>> pl.Config.set_tbl_cols(-1)
+            >>> pl.Config.set_tbl_rows(-1)
+            >>> df = pl.LazyFrame(
+            ...     dict(
+            ...         group=(1, 2, 1, 2),
+            ...         item_id=(1, 2, 2, 3),
+            ...         item_name=("a", "a", "b", "b"),
+            ...         value1=(1, np.nan, 3, 4),
+            ...         value2=range(4, 8),
+            ...     )
             ... )
-            >>> df
-            shape: (3, 3)
-            ┌───────┬────────────┬──────────────┐
-            │ Aloha ┆ Bell Chart ┆ Animals@#$%^ │
-            │ ---   ┆ ---        ┆ ---          │
-            │ i64   ┆ i64        ┆ i64          │
-            ╞═══════╪════════════╪══════════════╡
-            │ 0     ┆ 0          ┆ 0            │
-            │ 1     ┆ 1          ┆ 1            │
-            │ 2     ┆ 2          ┆ 2            │
-            └───────┴────────────┴──────────────┘
-            >>> df.janitor.clean_names(remove_special=True)
-            shape: (3, 3)
-            ┌───────┬────────────┬─────────┐
-            │ aloha ┆ bell_chart ┆ animals │
-            │ ---   ┆ ---        ┆ ---     │
-            │ i64   ┆ i64        ┆ i64     │
-            ╞═══════╪════════════╪═════════╡
-            │ 0     ┆ 0          ┆ 0       │
-            │ 1     ┆ 1          ┆ 1       │
-            │ 2     ┆ 2          ┆ 2       │
-            └───────┴────────────┴─────────┘
+            >>> df.collect()
+            shape: (4, 5)
+            ┌───────┬─────────┬───────────┬────────┬────────┐
+            │ group ┆ item_id ┆ item_name ┆ value1 ┆ value2 │
+            │ ---   ┆ ---     ┆ ---       ┆ ---    ┆ ---    │
+            │ i64   ┆ i64     ┆ str       ┆ i64    ┆ i64    │
+            ╞═══════╪═════════╪═══════════╪════════╪════════╡
+            │ 1     ┆ 1       ┆ a         ┆ 1      ┆ 4      │
+            │ 2     ┆ 2       ┆ a         ┆ null   ┆ 5      │
+            │ 1     ┆ 2       ┆ b         ┆ 3      ┆ 6      │
+            │ 2     ┆ 3       ┆ b         ┆ 4      ┆ 7      │
+            └───────┴─────────┴───────────┴────────┴────────┘
+
+            Generate all possible combinations of
+            `group`, `item_id`, and `item_name`
+            (whether or not they appear in the data)
+            >>> df.janitor.complete("group", "item_id", "item_name", sort=True).collect()
+            shape: (12, 5)
+            ┌───────┬─────────┬───────────┬────────┬────────┐
+            │ group ┆ item_id ┆ item_name ┆ value1 ┆ value2 │
+            │ ---   ┆ ---     ┆ ---       ┆ ---    ┆ ---    │
+            │ i64   ┆ i64     ┆ str       ┆ i64    ┆ i64    │
+            ╞═══════╪═════════╪═══════════╪════════╪════════╡
+            │ 1     ┆ 1       ┆ a         ┆ 1      ┆ 4      │
+            │ 1     ┆ 1       ┆ b         ┆ null   ┆ null   │
+            │ 1     ┆ 2       ┆ a         ┆ null   ┆ null   │
+            │ 1     ┆ 2       ┆ b         ┆ 3      ┆ 6      │
+            │ 1     ┆ 3       ┆ a         ┆ null   ┆ null   │
+            │ 1     ┆ 3       ┆ b         ┆ null   ┆ null   │
+            │ 2     ┆ 1       ┆ a         ┆ null   ┆ null   │
+            │ 2     ┆ 1       ┆ b         ┆ null   ┆ null   │
+            │ 2     ┆ 2       ┆ a         ┆ null   ┆ 5      │
+            │ 2     ┆ 2       ┆ b         ┆ null   ┆ null   │
+            │ 2     ┆ 3       ┆ a         ┆ null   ┆ null   │
+            │ 2     ┆ 3       ┆ b         ┆ 4      ┆ 7      │
+            └───────┴─────────┴───────────┴────────┴────────┘
+
+            Cross all possible `group` values with the unique pairs of
+            `(item_id, item_name)` that already exist in the data.
+            For such situations, where there is a group of columns,
+            pass it in as a struct:
+            >>> df.janitor.complete("group", pl.struct("item_id", "item_name"), sort=True).collect()
+            shape: (8, 5)
+            ┌───────┬─────────┬───────────┬────────┬────────┐
+            │ group ┆ item_id ┆ item_name ┆ value1 ┆ value2 │
+            │ ---   ┆ ---     ┆ ---       ┆ ---    ┆ ---    │
+            │ i64   ┆ i64     ┆ str       ┆ i64    ┆ i64    │
+            ╞═══════╪═════════╪═══════════╪════════╪════════╡
+            │ 1     ┆ 1       ┆ a         ┆ 1      ┆ 4      │
+            │ 1     ┆ 2       ┆ a         ┆ null   ┆ null   │
+            │ 1     ┆ 2       ┆ b         ┆ 3      ┆ 6      │
+            │ 1     ┆ 3       ┆ b         ┆ null   ┆ null   │
+            │ 2     ┆ 1       ┆ a         ┆ null   ┆ null   │
+            │ 2     ┆ 2       ┆ a         ┆ null   ┆ 5      │
+            │ 2     ┆ 2       ┆ b         ┆ null   ┆ null   │
+            │ 2     ┆ 3       ┆ b         ┆ 4      ┆ 7      │
+            └───────┴─────────┴───────────┴────────┴────────┘
+
+            Fill in nulls:
+            >>> df.janitor.complete(
+            ...     "group",
+            ...     pl.struct("item_id", "item_name"),
+            ...     fill_value={"value1": 0, "value2": 99},
+            ...     explicit=True,
+            ...     sort=True,
+            ... ).collect()
+            shape: (8, 5)
+            ┌───────┬─────────┬───────────┬────────┬────────┐
+            │ group ┆ item_id ┆ item_name ┆ value1 ┆ value2 │
+            │ ---   ┆ ---     ┆ ---       ┆ ---    ┆ ---    │
+            │ i64   ┆ i64     ┆ str       ┆ i64    ┆ i64    │
+            ╞═══════╪═════════╪═══════════╪════════╪════════╡
+            │ 1     ┆ 1       ┆ a         ┆ 1      ┆ 4      │
+            │ 1     ┆ 2       ┆ a         ┆ 0      ┆ 99     │
+            │ 1     ┆ 2       ┆ b         ┆ 3      ┆ 6      │
+            │ 1     ┆ 3       ┆ b         ┆ 0      ┆ 99     │
+            │ 2     ┆ 1       ┆ a         ┆ 0      ┆ 99     │
+            │ 2     ┆ 2       ┆ a         ┆ 0      ┆ 5      │
+            │ 2     ┆ 2       ┆ b         ┆ 0      ┆ 99     │
+            │ 2     ┆ 3       ┆ b         ┆ 4      ┆ 7      │
+            └───────┴─────────┴───────────┴────────┴────────┘
+
+            Limit the fill to only the newly created
+            missing values with `explicit = FALSE`
+            >>> df.janitor.complete(
+            ...     "group",
+            ...     pl.struct("item_id", "item_name"),
+            ...     fill_value={"value1": 0, "value2": 99},
+            ...     explicit=False,
+            ...     sort=True,
+            ... ).collect()
+            shape: (8, 5)
+            ┌───────┬─────────┬───────────┬────────┬────────┐
+            │ group ┆ item_id ┆ item_name ┆ value1 ┆ value2 │
+            │ ---   ┆ ---     ┆ ---       ┆ ---    ┆ ---    │
+            │ i64   ┆ i64     ┆ str       ┆ i64    ┆ i64    │
+            ╞═══════╪═════════╪═══════════╪════════╪════════╡
+            │ 1     ┆ 1       ┆ a         ┆ 1      ┆ 4      │
+            │ 1     ┆ 2       ┆ a         ┆ 0      ┆ 99     │
+            │ 1     ┆ 2       ┆ b         ┆ 3      ┆ 6      │
+            │ 1     ┆ 3       ┆ b         ┆ 0      ┆ 99     │
+            │ 2     ┆ 1       ┆ a         ┆ 0      ┆ 99     │
+            │ 2     ┆ 2       ┆ a         ┆ null   ┆ 5      │
+            │ 2     ┆ 2       ┆ b         ┆ 0      ┆ 99     │
+            │ 2     ┆ 3       ┆ b         ┆ 4      ┆ 7      │
+            └───────┴─────────┴───────────┴────────┴────────┘
 
         !!! info "New in version 0.28.0"
 
         Args:
             *columns: This refers to the columns to be completed.
-                It could be column labels (string type),
-                a list/tuple of column labels, or a dictionary that pairs
-                column labels with new values.
-            fill_value: Scalar value to use instead of nulls
+                It can be a string or a column selector or a polars expression.
+                A polars expression can be used to introduced new values,
+                as long as the polars expression has a name that already exists
+                in the LazyFrame.
+            fill_value: Scalar value or polars expression to use instead of nulls
                 for missing combinations. A dictionary, mapping columns names
                 to a scalar value is also accepted.
             explicit: Determines if only implicitly missing values
-                should be filled (`False`), or all nulls existing in the dataframe
+                should be filled (`False`), or all nulls existing in the LazyFrame
                 (`True`). `explicit` is applicable only
                 if `fill_value` is not `None`.
-            sort: Sort DataFrame based on *columns.
+            sort: Sort the LazyFrame based on *columns.
+            by: Column(s) to group by.
+                The explicit missing rows are returned per group.
         Returns:
-            A polars DataFrame.
+            A polars LazyFrame.
         """  # noqa: E501
         return _complete(
             df=self._df,
@@ -184,4 +422,5 @@ class PolarsLazyFrame:
             fill_value=fill_value,
             explicit=explicit,
             sort=sort,
+            by=by,
         )